+++ src/main/java/com/example/embedika_test/dao/dto/CarDto.java
@@ -11,7 +11,7 @@ public class CarDto {
     private final String carNumber;
     private final String color;
     private final String year;
-    private final Byte amountOfOwners;
+    private final Integer amountOfOwners;
     private Integer mileage;
     private CarBodyType bodyType;
     private final String regionNumber;
